requirements.txt
Django==3.1.7
django-debug-toolbar==3.2
